+++ CHANGES.md
@@ -12,6 +12,7 @@
    * DATA SOURCE panel
    * OPERATIONS panel
    * Variables selection dialog
+5. In **DATA SOURCES** and **WORKSPACE** panel: in details table show long text in popover
 6. Improved the default placeholder text for geometry fields